implemented tests for ByteArr...

+++ src/htcore/scripts/bytearr_read.cpp
@@ -44,7 +44,7 @@ ByteArrRead::ByteArrRead(const std::vector<uint8_t> &data) :
 
 /* public       */
 
-int ByteArrRead::getU08(unsigned int index)
+long ByteArrRead::getU08(unsigned int index)
 {
    uint8_t ret;
 
@@ -58,7 +58,7 @@ int ByteArrRead::getU08(unsigned int index)
    return ret;
 }
 
-int ByteArrRead::getU16(unsigned int index, int/*Endianness*/ end)
+long ByteArrRead::getU16(unsigned int index, int/*Endianness*/ end)
 {
    uint16_t ret;
 
@@ -81,7 +81,7 @@ int ByteArrRead::getU16(unsigned int index, int/*Endianness*/ end)
    return ret;
 }
 
-int ByteArrRead::getU32(unsigned int index, int/*Endianness*/ end)
+long ByteArrRead::getU32(unsigned int index, int/*Endianness*/ end)
 {
    uint32_t ret;
 
@@ -108,7 +108,7 @@ int ByteArrRead::getU32(unsigned int index, int/*Endianness*/ end)
    return ret;
 }
 
-int ByteArrRead::getS08(unsigned int index)
+long ByteArrRead::getS08(unsigned int index)
 {
    int8_t ret;
 
@@ -122,7 +122,7 @@ int ByteArrRead::getS08(unsigned int index)
    return ret;
 }
 
-int ByteArrRead::getS16(unsigned int index, int/*Endianness*/ end)
+long ByteArrRead::getS16(unsigned int index, int/*Endianness*/ end)
 {
    int16_t ret;
 
@@ -145,7 +145,7 @@ int ByteArrRead::getS16(unsigned int index, int/*Endianness*/ end)
    return ret;
 }
 
-int ByteArrRead::getS32(unsigned int index, int/*Endianness*/ end)
+long ByteArrRead::getS32(unsigned int index, int/*Endianness*/ end)
 {
    int32_t ret;
